level_editor folder renamed to Editor
