+++ src/Editor/editor_mode.c
