+++ src/Editor/text.c
