+++ src/Editor/text.h
