+++ src/Engine/files_parser.c
@@ -3,7 +3,7 @@
 #define JSMN_HEADER
 #include "../third_party/jsmn.h"
 #include "../engine.h"
-#include "../level_editor/data.h"
+#include "../Editor/data.h"
 
 typedef jsmntok_t Token;
