+++ src/main.c
@@ -3,7 +3,6 @@
 #include <EGL/egl.h>
 #include <GLES2/gl2.h>
 
-#include <vec3.h>
 #include "third_party/cglm/cglm.h"
 
 #include "utils.h"
